+++ website/api/moderation/addUserShadowBan.ts
@@ -2,16 +2,16 @@ import axios from "axios";
 
 import apiBaseUrl from "../../utils/apiBaseUrl";
 
-export default function removeUserBan(username, callback) {
+export default function addUserShadowBan(username: string, callback) {
     axios
         .put(
-            apiBaseUrl + "/moderation/remove-user-ban",
+            apiBaseUrl + "/moderation/add-user-shadow-ban",
             {
                 username: username,
             },
             {
                 withCredentials: true,
-            },
+            }
         )
         .then(function (response) {
             callback(response.data);
