+++ website/api/moderation/killComment.ts
@@ -2,16 +2,16 @@ import axios from "axios";
 
 import apiBaseUrl from "../../utils/apiBaseUrl";
 
-export default function addUserShadowBan(username, callback) {
+export default function killComment(commentId, callback) {
     axios
         .put(
-            apiBaseUrl + "/moderation/add-user-shadow-ban",
+            apiBaseUrl + "/moderation/kill-comment",
             {
-                username: username,
+                id: commentId,
             },
             {
                 withCredentials: true,
-            },
+            }
         )
         .then(function (response) {
             callback(response.data);
