+++ website/components/item.js
@@ -9,7 +9,7 @@ import unfavoriteItem from "../api/items/unfavoriteItem.js";
 import hideItem from "../api/items/hideItem.js";
 import unhideItem from "../api/items/unhideItem.js";
 import addNewComment from "../api/comments/addNewComment.js";
-import killItem from "../api/moderation/killItem.js";
+import killItem from "../api/moderation/killItem";
 import unkillItem from "../api/moderation/unkillItem.js";
 
 import renderCreatedTime from "../utils/renderCreatedTime";
